requirements/common.txt
onepassword-local-search==0.12.0
Click==7.0
click-log==0.3.2
PyYAML==5.1
pexpect==4.7.0
plumbum==1.6.7
